requirements.txt
python-memcached
nltk
swifter
PySastrawi
scikit-learn
aiohttp==3.8.4
aiosignal==1.3.1
aiostream==0.4.5
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.6
Django==3.2.20
django-crispy-forms==1.14.0
django-utils-six==2.0
fastapi==0.95.2
frozenlist==1.3.3
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
Pillow==9.5.0
protobuf==4.23.2
pydantic==1.10.8
Pygments==2.15.1
python-aiml==0.9.3
requests==2.31.0
rich==13.3.5
sigtools==4.0.1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
synchronicity==0.5.3
tblib==1.7.0
toml==0.10.2
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.6
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
watchfiles==0.19.0
yarl==1.9.2
zipp==3.15.0
